configure: add test for xz-utils
Some upstream now release tarballs compressed with xz, along with the
more traditional bzip2 and/or gzip tarballs.  xz (and its predecessor
lzma) achieves better compression, so is a good candidate to shorten
the time it takes to download the source tarballs.

Enable downloading xz-cmopressed tarballs if the user has xz-utils.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@anciens.enib.fr>