kernel/linux: deprecate copying headers from pre-instaleld headers
No-one ever inquired about this feature. Due to the way it works, it is
rather unlikely that every user used a properly-formatted tarball, or a
properly structured directory as imput.

Beside, it is not really easy to maintain: I just spent one full day
remembering how it worked... Sigh... :-(

Just deprecate this, until next release, or until someone complains.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@anciens.enib.fr>