samples: do not use the mirror by default
As the mirror is always tried first, and the default is my server,
it means every download will first hit my server. As I do not have
much mirrored there, it means a lot of time is lost.

Disable use of the mirror for all samples, except the one that does
need it for the uClibc snapshot.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@anciens.enib.fr>