tsar: add new sample for building toolchain
Our sample builds mipsel-tsar-linux-uclibc. It relies on a custom version of
linux and uclibc. For uclibc, a custom config file is provided as well. And
also, our sample patches gcc4.8.2, because even if we build gcc for the mips
ISA, we don't support everything: in this particular case, we don't support
unwinding.