tsar: deactivate default OPTSPACE option
Even though, gcc specifies that "branch likely" instructions should not be used
when compiling for mips32 arch, the "-Os" option seems to override that rule.
Disabling OPTSPACE option removes "-Os" which prevents from "-mbranch-likely"
to be activated.

Note that another way of doing that, if one want to keep the "-Os" option,
would be to explicitely specify "-mno-branch-likely" when compiling gcc's
libraries.